fix tests

+++ src/benchmark.c
@@ -1,4 +1,8 @@
 #define BENCHMARK
+#include <stdio.h>
+#include <string.h>
+#include <assert.h>
+
 #include "calculations.h"
 #include "config.h"
 #include "decoder.h"
@@ -9,8 +13,6 @@
 #include "table.h"
 #include "tasks.h"
 #include "util.h"
-#include <stdio.h>
-#include <string.h>
 
 static void do_fuel_calculation_bench() {
   platform_load_config();
@@ -18,7 +20,7 @@ static void do_fuel_calculation_bench() {
   uint64_t start = cycle_count();
   calculate_fueling();
   uint64_t end = cycle_count();
-  printf("calculate_fueling: %llu ns\r\n", cycles_to_ns(end - start));
+  printf("calculate_fueling: %llu ns\r\n", (unsigned long long)cycles_to_ns(end - start));
 }
 
 static void do_schedule_ignition_event_bench() {
@@ -40,21 +42,21 @@ static void do_schedule_ignition_event_bench() {
   assert(config.events[0].start.scheduled);
 
   printf("fresh ignition schedule_event: %llu ns\r\n",
-         cycles_to_ns(end - start));
+         (unsigned long long)cycles_to_ns(end - start));
 
   calculated_values.timing_advance = 25.0f;
   start = cycle_count();
   schedule_event(&config.events[0]);
   end = cycle_count();
 
-  printf("backward schedule_event: %llu ns\r\n", cycles_to_ns(end - start));
+  printf("backward schedule_event: %llu ns\r\n", (unsigned long long)cycles_to_ns(end - start));
 
   calculated_values.timing_advance = 15.0f;
   start = cycle_count();
   schedule_event(&config.events[0]);
   end = cycle_count();
 
-  printf("forward schedule_event: %llu ns\r\n", cycles_to_ns(end - start));
+  printf("forward schedule_event: %llu ns\r\n", (unsigned long long)cycles_to_ns(end - start));
 }
 
 static void do_sensor_adc_calcs() {
@@ -66,7 +68,7 @@ static void do_sensor_adc_calcs() {
   sensors_process(SENSOR_ADC);
   uint64_t end = cycle_count();
 
-  printf("process_sensor(adc): %llu ns\r\n", cycles_to_ns(end - start));
+  printf("process_sensor(adc): %llu ns\r\n", (unsigned long long)cycles_to_ns(end - start));
 }
 
 void do_buffer_swap_bench_best_case() {
@@ -76,7 +78,7 @@ void do_buffer_swap_bench_best_case() {
   scheduler_buffer_swap();
   uint64_t end = cycle_count();
 
-  printf("best case buffer_swap: %llu ns\r\n", cycles_to_ns(end - start));
+  printf("best case buffer_swap: %llu ns\r\n", (unsigned long long)cycles_to_ns(end - start));
 }
 
 void do_buffer_swap_bench_all_fired() {
@@ -88,7 +90,7 @@ void do_buffer_swap_bench_all_fired() {
   uint64_t end = cycle_count();
 
   printf("worst case 'fired' buffer_swap: %llu ns\r\n",
-         cycles_to_ns(end - start));
+         (unsigned long long)cycles_to_ns(end - start));
 }
 
 void do_buffer_swap_bench_all_ready() {
@@ -100,7 +102,7 @@ void do_buffer_swap_bench_all_ready() {
   uint64_t end = cycle_count();
 
   printf("worst case 'ready' buffer_swap: %llu ns\r\n",
-         cycles_to_ns(end - start));
+         (unsigned long long)cycles_to_ns(end - start));
 }
 
 int main() {
